Separate exclusion and fuzzier tests
