+++ src/test/kotlin/com/mituuz/fuzzier/FuzzierTest.kt
@@ -1,17 +1,8 @@
 package com.mituuz.fuzzier
 
-import com.intellij.openapi.components.service
-import com.intellij.openapi.project.DumbService
-import com.intellij.openapi.roots.ProjectFileIndex
-import com.intellij.psi.PsiDocumentManager
-import com.intellij.testFramework.PsiTestUtil
 import com.intellij.testFramework.TestApplicationManager
-import com.intellij.testFramework.fixtures.IdeaTestFixtureFactory
-import com.intellij.testFramework.runInEdtAndWait
-import com.mituuz.fuzzier.settings.FuzzierSettingsService
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
-import javax.swing.DefaultListModel
 
 class FuzzierTest {
     private var fuzzier: Fuzzier
@@ -20,91 +11,6 @@ class FuzzierTest {
     init {
         testApplicationManager = TestApplicationManager.getInstance()
         fuzzier = Fuzzier()
-    }
-
-    @Test
-    fun excludeListTest() {
-        service<FuzzierSettingsService>().state.exclusionList = listOf("asd", "nope")
-        val filePaths = listOf("src/main.kt", "src/asd/main.kt", "src/asd/asd.kt", "src/not/asd.kt", "src/nope")
-        val filePathContainer = setUpProjectFileIndex(filePaths)
-        assertEquals(1, filePathContainer.size())
-        assertEquals("/main.kt", filePathContainer.get(0).string)
-    }
-
-    @Test
-    fun excludeListTestNoMatches() {
-        service<FuzzierSettingsService>().state.exclusionList = listOf("asd")
-        val filePaths = listOf("src/main.kt", "src/not.kt", "src/dsa/not.kt")
-        val filePathContainer = setUpProjectFileIndex(filePaths)
-        assertEquals(3, filePathContainer.size())
-        assertEquals("/main.kt", filePathContainer.get(2).string)
-        assertEquals("/not.kt", filePathContainer.get(1).string)
-        assertEquals("/dsa/not.kt", filePathContainer.get(0).string)
-       }
-
-    @Test
-    fun excludeListTestEmptyList() {
-        service<FuzzierSettingsService>().state.exclusionList = ArrayList()
-        val filePaths = listOf("src/main.kt", "src/not.kt", "src/dsa/not.kt")
-        val filePathContainer = setUpProjectFileIndex(filePaths)
-        assertEquals(3, filePathContainer.size())
-        assertEquals("/main.kt", filePathContainer.get(2).string)
-        assertEquals("/not.kt", filePathContainer.get(1).string)
-        assertEquals("/dsa/not.kt", filePathContainer.get(0).string)
-    }
-
-    @Test
-    fun excludeListTestStartsWith() {
-        service<FuzzierSettingsService>().state.exclusionList = listOf("/asd*")
-        val filePaths = listOf("src/main.kt", "src/asd/main.kt", "src/asd/asd.kt", "src/not/asd.kt")
-        val filePathContainer = setUpProjectFileIndex(filePaths)
-        assertEquals(2, filePathContainer.size())
-        assertEquals("/not/asd.kt", filePathContainer.get(0).string)
-    }
-
-    @Test
-    fun excludeListTestEndsWith() {
-        service<FuzzierSettingsService>().state.exclusionList = listOf("*.log")
-        val filePaths = listOf("src/main.log", "src/asd/main.log", "src/asd/asd.kt", "src/not/asd.kt", "src/nope")
-        val filePathContainer = setUpProjectFileIndex(filePaths)
-        assertEquals(3, filePathContainer.size())
-        assertEquals("/asd/asd.kt", filePathContainer.get(0).string)
-    }
-
-    private fun setUpProjectFileIndex(filesToAdd: List<String>) : DefaultListModel<Fuzzier.FuzzyMatchContainer> {
-        val filePathContainer = DefaultListModel<Fuzzier.FuzzyMatchContainer>()
-
-        val factory = IdeaTestFixtureFactory.getFixtureFactory()
-        val fixtureBuilder = factory.createLightFixtureBuilder(null, "Test")
-        val fixture = fixtureBuilder.fixture
-
-        val myFixture = IdeaTestFixtureFactory.getFixtureFactory().createCodeInsightFixture(fixture)
-        myFixture.setUp()
-
-        filesToAdd.forEach {
-            myFixture.addFileToProject(it, "")
-        }
-
-        // Add source and wait for indexing
-        val dir = myFixture.findFileInTempDir("src")
-        val basePath = dir.canonicalPath
-        PsiTestUtil.addSourceRoot(fixture.module, dir)
-        runInEdtAndWait {
-            PsiDocumentManager.getInstance(fixture.project).commitAllDocuments()
-        }
-        DumbService.getInstance(fixture.project).waitForSmartMode()
-
-        val contentIterator = basePath?.let { fuzzier.getContentIterator(it, "", filePathContainer) }
-        val index = ProjectFileIndex.getInstance(fixture.project)
-        runInEdtAndWait {
-            if (contentIterator != null) {
-                index.iterateContent(contentIterator)
-            }
-        }
-        // Handle clearing ProjectFileIndex between tests
-        myFixture.tearDown()
-
-        return filePathContainer
     }
 
     @Test
