+++ src/test/kotlin/com/mituuz/fuzzier/TestUtil.kt
@@ -1,0 +1,24 @@
+package com.mituuz.fuzzier
+
+import com.intellij.openapi.project.DumbService
+import com.intellij.psi.PsiDocumentManager
+import com.intellij.testFramework.PsiTestUtil
+import com.intellij.testFramework.fixtures.CodeInsightTestFixture
+import com.intellij.testFramework.fixtures.IdeaProjectTestFixture
+import com.intellij.testFramework.runInEdtAndWait
+
+class TestUtil {
+    fun addFilesToProject(filesToAdd: List<String>, myFixture: CodeInsightTestFixture, fixture: IdeaProjectTestFixture) {
+        filesToAdd.forEach {
+            myFixture.addFileToProject(it, "")
+        }
+
+        // Add source and wait for indexing
+        val dir = myFixture.findFileInTempDir("src")
+        PsiTestUtil.addSourceRoot(fixture.module, dir)
+        runInEdtAndWait {
+            PsiDocumentManager.getInstance(fixture.project).commitAllDocuments()
+        }
+        DumbService.getInstance(fixture.project).waitForSmartMode()
+    }
+}
